07-项目实战/agent-agentic-rag/requirements.txt
# 向量模型
sentence-transformers>=3.0.0

# 检索与索引
rank-bm25>=0.2.2
faiss-cpu>=1.8.0
jieba>=0.42.1

# Agent 编排
langgraph>=0.2.0
